Allow `rev-list` to ignore missing objects
When calculating sets of commits to send between repositories, some of
the inputs will be IDs for commits that don't exist in the local repo.
In particular, excluded inputs that come from commits the remote peer
says it has might not exist in the local repository.

For example, consider the following history. Alice and Bob both have
copies of A and B. Alice has extended B with commits C and D, and Bob
has added E and F.

    A   B   C   D
    o---o---o---o [alice]
         \
          o---o [bob]
          E   F

If Bob wants to fetch from Alice, he'll say he wants D, and he has F.
That means he wants to receive the set of commits F..D. If Alice has E
and F, then she can compute the range and send only commits C and D. But
if she does not have these commits, we don't want the operation to fail.
If Alice ignores that she does not have F, then she'll compute the range
that includes D, giving commits A, B, C and D. This isn't *wrong* per
se, it's just inefficient, because she sends commits that Bob already
has. But this is better than sending nothing at all.

Therefore, when computing a `RevList` from IDs supplied by multiple
parties, we should ignore commits that are missing. The `:missing`
option to `RevList` enables this behaviour.