Add a `--reverse` option to `rev-list`
This is purely for the benefit of checking `RevList` output while we
develop the `fetch` and `push` commands. If `--reverse` is passed to
`rev-list`, it calls `reverse_each` rather than `each`. We don't need a
`:reverse` option on `RevList` itself since callers can just call
`reverse_each` if they want that behaviour.