Support seeking backwards in an incoming pack stream
When reading objects from a pack, there's no indication of how long the
zlib-compressed buffer for each object is -- the record header gives the
size before compression, not after. Since we don't know where the object
ends, the safest thing to do is feed input into zlib one byte at a time
and stop when zlib says the stream is finished.

    def read_zlib_stream
      stream = Zlib::Inflate.new
      string = ""

      until stream.finished?
        string.concat(stream.inflate(@input.read(1)))
      end

      string
    end

However, drip-feeding data into zlib like this is really slow, as a our
program spends most of its time in a Ruby loop calling `read` and
`finished?`. By reading in chunks we spend more time in zlib on each
iteration, but we might read past the end of the object. We need to
relocate the read cursor to the exact end of the object so we can begin
reading the next one. A method for doing that using the Ruby `IO`
interface might look like this:

    def read_zlib_stream
      stream = Zlib::Inflate.new
      string = ""
      total  = 0

      until stream.finished?
        data   = @input.read_nonblock(256)
        total += data.bytesize

        string.concat(stream.inflate(data))
      end
      @input.seek(stream.total_in - total, IO::SEEK_CUR)

      string
    end

However, streaming `IO` sources don't support `seek` -- once you've read
some data, the `IO` object forgets about it. We can recover this kind of
behaviour by implementing it on `Pack::Stream` which will wrap the
underlying `IO` stream.

By using the `capture` method, the caller can make `Stream` store up any
data that's read while some block is executed. `seek` can then use this
captured data to put some of it back in a buffer. The `read` and
`read_nonblock` methods pull from this buffer before fetching more data
from the underlying `IO`.

If capturing is enabled, then data is not passed into the SHA-1 check
until the capture session ends, as if we read past the end of an object
we might be reading the SHA-1 at the end of the pack. So, we only update
our digest once we know how much data actually contributed to the
current object.