Reader for the pack format
The `Pack::Reader` class inverts the work done by `Pack::Writer`. On the
receiving end of the connection, e.g. in the `fetch` command, it wraps a
readable `IO` stream and provides methods for reading the pack header
and then reading each record.

By using the `read_nonblock` and `seek` methods from `Pack::Stream`, it
reads the inputs in chunks but makes sure to put the read cursor at the
end of each object, ready to read the next one.

Reading an object from a pack returns a `Pack::Record`, rather than one
of the `Database` object types. That's because packs are more like a
low-level data store -- they just need to yield enough information about
the object to save it to a file in `.git/objects`, or to provide its raw
data so that other code can fully parse the object. `Pack::Record` has
an interface that means it can be passed to `Database#store` and be
written to disk.