Connect to remote repositories over SSH
The `fetch` command we just introduced only supports `file:` URLs. This
can be extended by changing `build_agent_command` to generate an SSH
command that runs `git-upload-pack` on the remote machine. For example,
if we have a remote configured:

    [remote "github"]
        url = ssh://git@github.com/jcoglan/jit.git

Then we should run this command to connect to it:

    ssh github.com -l git "git-upload-pack /jcoglan/jit.git"

This change in `RemoteClient` does this, and the rest of the protocol
continues just the same as if we were talking to a local repository.