Variable shift for variable-length integers
The pack format [1] uses a variable-length number encoding to represent
the size of each object. The first byte contains the four
least-significant bits, the second byte the next seven least-significant
bits, and so on.

Deltas use a very similar format to encode the lengths of the source and
target strings. The difference is that the first byte encodes seven
rather than four bits. We accommodate this by parameterising `VarIntLE`
on the number of bits in the first byte.

[1]: https://git-scm.com/docs/pack-format