Decoder for deltas
When receiving a pack containing delta records, we need to be able to
expand the delta to regenerate the original object. The `Pack::Expander`
class does this by parsing the delta into its instructions, and applying
these instructions to some source text.

It checks the given sizes of the source and target texts to make sure
the text we generate is consistent with what the compressor wrote.

One quirk in this format is that Git caps the `size` field of copy
instructions at 0x10000, even though the serialisation format can
include sizes up to 0xffffff. If a copy instruction has size 0x10000,
Git sends the value 0 instead to save space. So when decoding, we need
to check for copy instructions with zero size -- they don't mean copy
an empty chunk, they mean Git hit its size limit for that instruction.