Send delta-compressed objects in packs
Having built up the tools for delta compression over the last few
commits, we can integrate it into the `Writer` and `Reader` classes for
packs.

`Writer` uses the `Compressor` class to find deltas between all its
objects, before it writes them out into the pack. Every object must
appear before any objects that use it as a delta base, so `write_entry`
now needs to recursively write the base object of each entry, if that
object has not already been written. We track this by setting the
`offset` on each `Entry` when it's written to the stream.

The `Reader` class now needs to emit a different kind of value when it
reads a delta record. It should not try to reconstitute the original
object itself, since there will be multiple ways of reloading the base
object data, depending on where the `Reader` is being used. Instead it
returns a `RefDelta` containing the base object ID and delta data.

Then, we use a new class called `Unpacker` to sit between `Pack::Reader`
and the database. When receiving a pack, it writes each object to the
database, and for delta records, it reloads the base object from the DB
and uses it to expand the delta. Due to the ordering of objects within
packs, the base object should be in the database by the time it's
needed.