Keep packs as primary storage
Git has three variables that set an "unpack limit" when receiving a pack
from a remote repo: `fetch.unpackLimit` and `receive.unpackLimit` set
the limits for the `fetch` and `receive-pack` commands, and both fall
back to `transfer.unpackLimit` if not set.

If these variables are set, and the incoming pack contains more objects
than the limit specifies, then we'll use `Pack::Indexer` to store the
pack on disk, instead of `Pack::Unpacker` to extract its contents.

We don't yet have code for using pack files as sources for the
`Database`, so we can't yet read objects back out of them. Therefore we
won't turn this behaviour on by default -- an `unpackLimit` variable
must be set to trigger it. (`transfer.unpackLimit` usually defaults to
100 if not set.)