Set an upstream for a branch
To make it easier to fetch and pull content between repos, Git allows
local branches to be "linked" with remote-tracking ones using the
`--set-upstream-to` option with `branch`. For example, if we have this
remote:

    [remote "origin"]
        fetch = +refs/heads/*:refs/remotes/origin/bar

And then we run this command:

    $ jit branch foo --set-upstream-to origin/bar

Then the following should be added to the config:

    [branch "foo"]
        remote = origin
        merge = refs/heads/bar

This tells us the name of the remote and branch that our local branch
should pull from and push to.

The `merge` variable is inferred by reversing the `remote.origin.fetch`
refspec: `origin/bar` expands to the ref `refs/remotes/origin/bar`,
which the refspec tells us is a copy of `refs/heads/bar` in the `origin`
remote.

The `--track` option can be used to set up this link when a branch is
first created. The `--unset-upstream` option removes the link.